refine more buffer code
