+++ java/fury-core/src/main/java/org/apache/fury/memory/MemoryUtils.java
@@ -67,11 +67,6 @@ public class MemoryUtils {
     }
   }
 
-  /** Get short in big endian order from provided buffer. */
-  public static short getShortB(byte[] b, int off) {
-    return (short) ((b[off + 1] & 0xFF) + (b[off] << 8));
-  }
-
   // Lazy load offset and also follow graalvm offset auto replace pattern.
   private static class Offset {
     private static final long BAS_BUF_BUF;
@@ -135,71 +130,4 @@ public class MemoryUtils {
     buffer.readerIndex(pos);
   }
 
-  public static int putVarUint36Small(byte[] arr, int index, long v) {
-    if (v >>> 7 == 0) {
-      arr[index] = (byte) v;
-      return 1;
-    }
-    if (v >>> 14 == 0) {
-      arr[index++] = (byte) ((v & 0x7F) | 0x80);
-      arr[index] = (byte) (v >>> 7);
-      return 2;
-    }
-    return bigWriteUint36(arr, index, v);
-  }
-
-  private static int bigWriteUint36(byte[] arr, int index, long v) {
-    if (v >>> 21 == 0) {
-      arr[index++] = (byte) ((v & 0x7F) | 0x80);
-      arr[index++] = (byte) (v >>> 7 | 0x80);
-      arr[index] = (byte) (v >>> 14);
-      return 3;
-    }
-    if (v >>> 28 == 0) {
-      arr[index++] = (byte) ((v & 0x7F) | 0x80);
-      arr[index++] = (byte) (v >>> 7 | 0x80);
-      arr[index++] = (byte) (v >>> 14 | 0x80);
-      arr[index] = (byte) (v >>> 21);
-      return 4;
-    }
-    arr[index++] = (byte) ((v & 0x7F) | 0x80);
-    arr[index++] = (byte) (v >>> 7 | 0x80);
-    arr[index++] = (byte) (v >>> 14 | 0x80);
-    arr[index++] = (byte) (v >>> 21 | 0x80);
-    arr[index] = (byte) (v >>> 28);
-    return 5;
-  }
-
-  public static void putInt(Object o, long pos, int value) {
-    if (!Platform.IS_LITTLE_ENDIAN) {
-      value = Integer.reverseBytes(value);
-    }
-    Platform.putInt(o, pos, value);
-  }
-
-  public static int getInt(Object o, long pos) {
-    int i = Platform.getInt(o, pos);
-    return Platform.IS_LITTLE_ENDIAN ? i : Integer.reverseBytes(i);
-  }
-
-  public static long getLong(Object o, long pos) {
-    long v = Platform.getLong(o, pos);
-    return Platform.IS_LITTLE_ENDIAN ? v : Long.reverseBytes(v);
-  }
-
-  public static void putFloat(Object o, long pos, float value) {
-    int v = Float.floatToRawIntBits(value);
-    if (!Platform.IS_LITTLE_ENDIAN) {
-      v = Integer.reverseBytes(v);
-    }
-    Platform.putInt(o, pos, v);
-  }
-
-  public static void putDouble(Object o, long pos, double value) {
-    long v = Double.doubleToRawLongBits(value);
-    if (!Platform.IS_LITTLE_ENDIAN) {
-      v = Long.reverseBytes(v);
-    }
-    Platform.putLong(o, pos, v);
-  }
 }
